+++ esibd/examples/custom_plugin/custom_plugin.py
@@ -13,17 +13,16 @@ def providePlugins() -> 'list[type[Plugin]]':
 
 
 class CustomPlugin(Plugin):
-    """The minimal code in "examples/Custom.py" demonstrates how to integrate your own custom elements to the ESIBD Explorer.
+    """The minimal code in "examples/custom_plugin/custom_plugin.py" demonstrates how to integrate your own custom elements to the ESIBD Explorer.
 
-    This should be sufficient as
-    long as your code does not requires interaction with any other elements
-    of the ESIBD Explorer. See :ref:`sec:plugin_system` for more information.
+    It also demonstrates how to interact with and even extend other plugins including internal plugins.
+    See :ref:`sec:plugin_system` for more information.
     """
 
-    documentation = """The minimal code in examples/Custom.py demonstrates how to integrate your own
-    custom elements to the ESIBD Explorer. This should be sufficient as
-    long as your code does not requires interaction with any other elements
-    of the ESIBD Explorer."""
+    documentation = """The minimal code in examples/custom_plugin/custom_plugin.py demonstrates how to integrate your own
+    custom elements to the ESIBD Explorer.
+    It also demonstrates how to interact with and even extend other plugins including internal plugins.
+    """
 
     name = 'CustomControl'
     version = '1.0'
@@ -55,3 +54,11 @@ class CustomPlugin(Plugin):
         lay.addWidget(lbl)
         dlg.setLayout(lay)
         dlg.exec()
+
+    def afterFinalizeInit(self) -> None:
+        super().afterFinalizeInit()
+        # NOTE: the next line demonstrated that you can even use a custom plugin to modify any of the build-in plugins.
+        # This may make debugging and maintaining the code harder, but it could also be cleaner and easier to implement
+        # compared to overwriting the internal plugin with a custom version.
+        self.pluginManager.Console.addAction(event=self.onClick, toolTip=f'Action added by {self.name}.', icon=self.makeIcon('cookie.png'),
+                                              before=self.pluginManager.Console.errorFilterAction)